RAG-Bot/requirements.txt
streamlit
langchain
langchain-community
langchain-cohere
langchain-groq
chromadb
cohere
pydantic
sqlite3
python-dotenv
tiktoken
PyPDF2